// src/app/models/usuario.interface.ts
export interface Usuario {
  userId?: number;
  userCuenta?: number;
  userTdId?: number;
  userNumeroDoc?: string;
  userNombre?: string;
  userApellido?: string;
  userTelefono?: string;
  userCorreo?: string;
  userPassword?: string;
  userdireccion?: string;
  userComprasRealizadas?: number;
  userVentasRealizadas?: number;
  userValoracionVenta?: number;
  userValoracionCompra?: number;
}